ctrlx_datalayer::client

Module bulk

Modules

async_bulk

creator_bulk

sync_bulk

Structs

BulkCreateArg

Input argument for a single entry in a bulk create request.

BulkReadArg

Input argument for a single entry in a bulk read request.

BulkResponse

The response for a single entry in a bulk operation.

BulkWriteArg

Input argument for a single entry in a bulk write request.

Traits

ResponseCallback

Callback trait for asynchronous bulk operation responses.

Type Aliases

ResponseCallbackBox

Type alias for a boxed ResponseCallback.