ctrlx_datalayer::system

Struct System

pub struct System { /* private fields */ }

Represents a ctrlX Data Layer system instance.

The System is the root object of the Data Layer. It manages the lifecycle of the underlying native system handle and provides access to the factory and converter subsystems.

let system = System::new("");
system.start(false);
{
  ...
}
info!("Shutdown!");

Implementations

impl System

pub fn new(path: &str) -> Self

Creates a new System instance.

Arguments

path - Path to the IPC endpoint used to connect to the Data Layer broker.

pub fn start(&self, start_broker: bool)

Starts the Data Layer system.

Arguments

start_broker - If true, the integrated broker is started along with the system.

pub fn stop(&self, force_provider_stop: bool) -> bool

Stops the Data Layer system.

Returns true if the system was stopped successfully.

Arguments

force_provider_stop - If true, all providers are forcefully stopped.

pub fn factory(&self) -> Factory

Returns the Factory associated with this system.

The factory is used to create clients and providers for the Data Layer.

pub fn converter(&self) -> Converter

Returns the Converter associated with this system.

The converter can be used to convert Data Layer variant values to and from JSON.

pub fn update_bfbs_path(&self, path: &str)

Sets the path to the BFBS (Binary FlatBuffers Schema) file.

Arguments

path - Path to the BFBS file.